Refactor ENSIndexer data model for Indexing Status
The goal of this change is to split deserialization from validation, so it's possible to: a) deserialize a serialized value, including validation; b) validate an unvalidated business-layer value.
