+++ packages/ensnode-sdk/src/api/indexing-status/zod-schemas.ts
@@ -1,12 +1,17 @@
 import { z } from "zod/v4";
 
-import { makeRealtimeIndexingStatusProjectionSchema } from "../../ensindexer/indexing-status/zod-schemas";
+import { makeSerializedRealtimeIndexingStatusProjectionSchema } from "../../ensindexer/indexing-status/deserialize/realtime-indexing-status-projection";
+import { makeRealtimeIndexingStatusProjectionSchema } from "../../ensindexer/indexing-status/validate/realtime-indexing-status-projection";
 import {
   type IndexingStatusResponse,
   IndexingStatusResponseCodes,
   type IndexingStatusResponseError,
   type IndexingStatusResponseOk,
 } from "./response";
+import {
+  SerializedIndexingStatusResponse,
+  SerializedIndexingStatusResponseOk,
+} from "./serialized-response";
 
 /**
  * Schema for {@link IndexingStatusResponseOk}
@@ -17,6 +22,17 @@ export const makeIndexingStatusResponseOkSchema = (
   z.strictObject({
     responseCode: z.literal(IndexingStatusResponseCodes.Ok),
     realtimeProjection: makeRealtimeIndexingStatusProjectionSchema(valueLabel),
+  });
+
+/**
+ * Schema for {@link SerializedIndexingStatusResponseOk}
+ **/
+export const makeSerializedIndexingStatusResponseOkSchema = (
+  valueLabel: string = "Serialized Indexing Status Response OK",
+) =>
+  z.strictObject({
+    responseCode: z.literal(IndexingStatusResponseCodes.Ok),
+    realtimeProjection: makeSerializedRealtimeIndexingStatusProjectionSchema(valueLabel),
   });
 
 /**
@@ -35,5 +51,16 @@ export const makeIndexingStatusResponseErrorSchema = (
 export const makeIndexingStatusResponseSchema = (valueLabel: string = "Indexing Status Response") =>
   z.discriminatedUnion("responseCode", [
     makeIndexingStatusResponseOkSchema(valueLabel),
+    makeIndexingStatusResponseErrorSchema(valueLabel),
+  ]);
+
+/**
+ * Schema for {@link SerializedIndexingStatusResponse}
+ **/
+export const makeSerializedIndexingStatusResponseSchema = (
+  valueLabel: string = "Serialized Indexing Status Response",
+) =>
+  z.discriminatedUnion("responseCode", [
+    makeSerializedIndexingStatusResponseOkSchema(valueLabel),
     makeIndexingStatusResponseErrorSchema(valueLabel),
   ]);
