+++ packages/ensnode-sdk/src/ensindexer/indexing-status/conversions.test.ts
@@ -1,18 +1,27 @@
 import { describe, expect, it } from "vitest";
 
-import { deserializeOmnichainIndexingStatusSnapshot } from "./deserialize";
-import { serializeOmnichainIndexingStatusSnapshot } from "./serialize";
-import type { SerializedOmnichainIndexingStatusSnapshot } from "./serialized-types";
-import { earlierBlockRef, earliestBlockRef, laterBlockRef, latestBlockRef } from "./test-helpers";
+import {
+  earlierBlockRef,
+  earliestBlockRef,
+  laterBlockRef,
+  latestBlockRef,
+} from "./block-refs.mock";
 import {
   ChainIndexingConfigTypeIds,
   ChainIndexingStatusIds,
   type ChainIndexingStatusSnapshotBackfill,
   type ChainIndexingStatusSnapshotFollowing,
   type ChainIndexingStatusSnapshotQueued,
+} from "./chain-indexing-status-snapshot";
+import { deserializeOmnichainIndexingStatusSnapshot } from "./deserialize/omnichain-indexing-status-snapshot";
+import {
   OmnichainIndexingStatusIds,
   type OmnichainIndexingStatusSnapshot,
-} from "./types";
+} from "./omnichain-indexing-status-snapshot";
+import {
+  type SerializedOmnichainIndexingStatusSnapshot,
+  serializeOmnichainIndexingStatusSnapshot,
+} from "./serialize/omnichain-indexing-status-snapshot";
 
 describe("ENSIndexer: Indexing Status", () => {
   describe("Omnichain Indexing Status Snapshot", () => {
