+++ packages/ensnode-sdk/src/ensindexer/indexing-status/omnichain-indexing-status-snapshot.test.ts
@@ -1,0 +1,205 @@
+import { describe, expect, it } from "vitest";
+
+import type { BlockRef } from "../../shared/types";
+import {
+  earlierBlockRef,
+  earliestBlockRef,
+  laterBlockRef,
+  latestBlockRef,
+} from "./block-refs.mock";
+import {
+  ChainIndexingConfigTypeIds,
+  ChainIndexingStatusIds,
+  type ChainIndexingStatusSnapshotBackfill,
+  type ChainIndexingStatusSnapshotCompleted,
+  type ChainIndexingStatusSnapshotFollowing,
+  type ChainIndexingStatusSnapshotQueued,
+} from "./chain-indexing-status-snapshot";
+import {
+  chainStatusesBackfillMixed,
+  chainStatusesCompleted,
+  chainStatusesFollowingMixed,
+  chainStatusesQueued,
+} from "./chain-indexing-status-snapshot.mocks";
+import {
+  getOmnichainIndexingCursor,
+  getOmnichainIndexingStatus,
+  OmnichainIndexingStatusIds,
+} from "./omnichain-indexing-status-snapshot";
+
+describe("Omnichain Indexing Status", () => {
+  describe("getOmnichainIndexingStatus", () => {
+    it("can correctly derive 'completed' status if all chains are 'completed'", () => {
+      // arrange
+      const chainStatuses = chainStatusesCompleted;
+
+      // act
+      const overallIndexingStatus = getOmnichainIndexingStatus(chainStatuses);
+
+      // assert
+      expect(overallIndexingStatus).toStrictEqual(OmnichainIndexingStatusIds.Completed);
+    });
+
+    it("can correctly derive 'unstarted' status if all chains are in 'queued' status", () => {
+      // arrange
+      const chainStatuses = chainStatusesQueued;
+
+      // act
+      const overallIndexingStatus = getOmnichainIndexingStatus(chainStatuses);
+
+      // assert
+      expect(overallIndexingStatus).toStrictEqual(OmnichainIndexingStatusIds.Unstarted);
+    });
+
+    it("can correctly derive 'backfill' status if all chains are either 'queued', 'backfill' or 'completed'", () => {
+      // arrange
+      const chainStatuses = chainStatusesBackfillMixed;
+
+      // act
+      const overallIndexingStatus = getOmnichainIndexingStatus(chainStatuses);
+
+      // assert
+      expect(overallIndexingStatus).toStrictEqual(OmnichainIndexingStatusIds.Backfill);
+    });
+
+    it("can correctly derive 'following' status if at least one chain is 'following", () => {
+      // arrange
+      const chainStatuses = chainStatusesFollowingMixed;
+
+      // act
+      const overallIndexingStatus = getOmnichainIndexingStatus(chainStatuses);
+
+      // assert
+      expect(overallIndexingStatus).toStrictEqual(OmnichainIndexingStatusIds.Following);
+    });
+  });
+
+  describe("getOmnichainIndexingCursor", () => {
+    it("returns the correct cursor for the given chains in any status", () => {
+      // arrange
+      const evenLaterBlockRef: BlockRef = {
+        timestamp: latestBlockRef.timestamp + 1000,
+        number: latestBlockRef.number + 1000,
+      };
+
+      const chainStatuses = [
+        {
+          chainStatus: ChainIndexingStatusIds.Queued,
+          config: {
+            configType: ChainIndexingConfigTypeIds.Indefinite,
+            startBlock: evenLaterBlockRef,
+          },
+        } satisfies ChainIndexingStatusSnapshotQueued,
+
+        {
+          chainStatus: ChainIndexingStatusIds.Backfill,
+          config: {
+            configType: ChainIndexingConfigTypeIds.Definite,
+            startBlock: earliestBlockRef,
+            endBlock: latestBlockRef,
+          },
+          latestIndexedBlock: earlierBlockRef,
+          backfillEndBlock: laterBlockRef,
+        } satisfies ChainIndexingStatusSnapshotBackfill,
+
+        {
+          chainStatus: ChainIndexingStatusIds.Following,
+          config: {
+            configType: ChainIndexingConfigTypeIds.Indefinite,
+            startBlock: earliestBlockRef,
+          },
+          latestIndexedBlock: earlierBlockRef,
+          latestKnownBlock: laterBlockRef,
+        } satisfies ChainIndexingStatusSnapshotFollowing,
+        {
+          chainStatus: ChainIndexingStatusIds.Completed,
+          config: {
+            configType: ChainIndexingConfigTypeIds.Definite,
+            startBlock: earlierBlockRef,
+            endBlock: latestBlockRef,
+          },
+          latestIndexedBlock: latestBlockRef,
+        } satisfies ChainIndexingStatusSnapshotCompleted,
+      ];
+
+      // act
+      const omnichainIndexingCursor = getOmnichainIndexingCursor(chainStatuses);
+
+      // assert
+      expect(omnichainIndexingCursor).toEqual(latestBlockRef.timestamp);
+    });
+
+    it("returns the correct cursor for the given queued chains only", () => {
+      expect(
+        getOmnichainIndexingCursor([
+          {
+            chainStatus: ChainIndexingStatusIds.Queued,
+            config: {
+              configType: ChainIndexingConfigTypeIds.Indefinite,
+              startBlock: earliestBlockRef,
+            },
+          } satisfies ChainIndexingStatusSnapshotQueued,
+          {
+            chainStatus: ChainIndexingStatusIds.Queued,
+            config: {
+              configType: ChainIndexingConfigTypeIds.Indefinite,
+              startBlock: laterBlockRef,
+            },
+          } satisfies ChainIndexingStatusSnapshotQueued,
+        ]),
+      ).toEqual(earliestBlockRef.timestamp - 1);
+    });
+
+    it("returns the correct cursor for the given indexed chains", () => {
+      // arrange
+      const evenLaterBlockRef: BlockRef = {
+        timestamp: latestBlockRef.timestamp + 1000,
+        number: latestBlockRef.number + 1000,
+      };
+
+      const chainStatuses = [
+        {
+          chainStatus: ChainIndexingStatusIds.Backfill,
+          config: {
+            configType: ChainIndexingConfigTypeIds.Definite,
+            startBlock: earliestBlockRef,
+            endBlock: latestBlockRef,
+          },
+          latestIndexedBlock: earlierBlockRef,
+          backfillEndBlock: laterBlockRef,
+        } satisfies ChainIndexingStatusSnapshotBackfill,
+
+        {
+          chainStatus: ChainIndexingStatusIds.Following,
+          config: {
+            configType: ChainIndexingConfigTypeIds.Indefinite,
+            startBlock: earliestBlockRef,
+          },
+          latestIndexedBlock: evenLaterBlockRef,
+          latestKnownBlock: laterBlockRef,
+        } satisfies ChainIndexingStatusSnapshotFollowing,
+        {
+          chainStatus: ChainIndexingStatusIds.Completed,
+          config: {
+            configType: ChainIndexingConfigTypeIds.Definite,
+            startBlock: earlierBlockRef,
+            endBlock: latestBlockRef,
+          },
+          latestIndexedBlock: latestBlockRef,
+        } satisfies ChainIndexingStatusSnapshotCompleted,
+      ];
+
+      // act
+      const omnichainIndexingCursor = getOmnichainIndexingCursor(chainStatuses);
+
+      // assert
+      expect(omnichainIndexingCursor).toEqual(evenLaterBlockRef.timestamp);
+    });
+
+    it("throws error when no chains were provided", () => {
+      expect(() => getOmnichainIndexingCursor([])).toThrowError(
+        /Unable to determine omnichain indexing cursor/,
+      );
+    });
+  });
+});
