+++ packages/ensnode-sdk/src/internal.ts
@@ -20,7 +20,6 @@ export * from "./api/shared/errors/zod-schemas";
 export * from "./api/shared/pagination/zod-schemas";
 export * from "./ensapi/config/zod-schemas";
 export * from "./ensindexer/config/zod-schemas";
-export * from "./ensindexer/indexing-status/zod-schemas";
 export * from "./registrars/zod-schemas";
 export * from "./rpc";
 export * from "./shared/config/build-rpc-urls";
